requirements.txt
openai
pandas
numpy
tiktoken
python-dotenv
tqdm